Remove raft_handler
