+++ src/handlers/mod.rs
@@ -1,6 +1,5 @@
 mod client_request_handler;
 mod fsck_handler;
-mod raft_handler;
 mod router;
 
 pub use client_request_handler::client_request_handler;
